* Refactor
  - Move file system related files to new FileSystem folder.
  - Move shaders resources to new folder.
  - Don't throw exception when file can't be open.
  - Add asserts and checks when file is not open.
  - Add combine and get current path functions in Path class.
  - Rename IO and Utility namespaced to FileSystem for file system related classes.
  - Change Debug config to include memory tests.
  - Add configuration for Debug with logging and no memory checks.
  - Add missing configurations.
