+++ K9ngine_v1.sln
@@ -20,6 +20,9 @@ Global
 		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		DebugLogNoMemory|Any CPU = DebugLogNoMemory|Any CPU
+		DebugLogNoMemory|x64 = DebugLogNoMemory|x64
+		DebugLogNoMemory|x86 = DebugLogNoMemory|x86
 		DebugMemoryNoLog|Any CPU = DebugMemoryNoLog|Any CPU
 		DebugMemoryNoLog|x64 = DebugMemoryNoLog|x64
 		DebugMemoryNoLog|x86 = DebugMemoryNoLog|x86
@@ -34,6 +37,12 @@ Global
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.Debug|x64.Build.0 = Debug|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.Debug|x86.Build.0 = Debug|Any CPU
+		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugLogNoMemory|Any CPU.ActiveCfg = DebugLogNoMemory|Any CPU
+		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugLogNoMemory|Any CPU.Build.0 = DebugLogNoMemory|Any CPU
+		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugLogNoMemory|x64.ActiveCfg = DebugLogNoMemory|Any CPU
+		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugLogNoMemory|x64.Build.0 = DebugLogNoMemory|Any CPU
+		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugLogNoMemory|x86.ActiveCfg = DebugLogNoMemory|Any CPU
+		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugLogNoMemory|x86.Build.0 = DebugLogNoMemory|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugMemoryNoLog|Any CPU.ActiveCfg = DebugNoLog|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugMemoryNoLog|Any CPU.Build.0 = DebugNoLog|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.DebugMemoryNoLog|x64.ActiveCfg = DebugMemoryNoLog|Any CPU
@@ -46,18 +55,24 @@ Global
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.Release|x64.Build.0 = Release|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.Release|x86.ActiveCfg = Release|Any CPU
 		{253EA720-1E9F-4155-B7D6-87ECD113A686}.Release|x86.Build.0 = Release|Any CPU
-		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|Any CPU.ActiveCfg = Debug|Win32
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|Any CPU.Build.0 = Debug|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|x64.ActiveCfg = Debug|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|x64.Build.0 = Debug|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|x86.ActiveCfg = Debug|Win32
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Debug|x86.Build.0 = Debug|Win32
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugLogNoMemory|Any CPU.ActiveCfg = DebugLogNoMemory|x64
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugLogNoMemory|Any CPU.Build.0 = DebugLogNoMemory|x64
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugLogNoMemory|x64.ActiveCfg = DebugLogNoMemory|x64
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugLogNoMemory|x86.ActiveCfg = DebugLogNoMemory|Win32
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugMemoryNoLog|Any CPU.ActiveCfg = DebugMemoryNoLog|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugMemoryNoLog|Any CPU.Build.0 = DebugMemoryNoLog|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugMemoryNoLog|x64.ActiveCfg = DebugMemoryNoLog|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugMemoryNoLog|x64.Build.0 = DebugMemoryNoLog|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugMemoryNoLog|x86.ActiveCfg = DebugMemoryNoLog|Win32
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.DebugMemoryNoLog|x86.Build.0 = DebugMemoryNoLog|Win32
-		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Release|Any CPU.ActiveCfg = Release|Win32
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Release|Any CPU.ActiveCfg = Release|x64
+		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Release|Any CPU.Build.0 = Release|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Release|x64.ActiveCfg = Release|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Release|x64.Build.0 = Release|x64
 		{C318273B-7DE1-4A22-AC9E-D3E4B99521AD}.Release|x86.ActiveCfg = Release|Win32
@@ -68,6 +83,12 @@ Global
 		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.Debug|x64.Build.0 = Debug|x64
 		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.Debug|x86.ActiveCfg = Debug|Win32
 		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.Debug|x86.Build.0 = Debug|Win32
+		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugLogNoMemory|Any CPU.ActiveCfg = DebugLogNoMemory|x64
+		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugLogNoMemory|Any CPU.Build.0 = DebugLogNoMemory|x64
+		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugLogNoMemory|x64.ActiveCfg = DebugLogNoMemory|x64
+		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugLogNoMemory|x64.Build.0 = DebugLogNoMemory|x64
+		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugLogNoMemory|x86.ActiveCfg = DebugLogNoMemory|Win32
+		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugLogNoMemory|x86.Build.0 = DebugLogNoMemory|Win32
 		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugMemoryNoLog|Any CPU.ActiveCfg = DebugMemoryNoLog|x64
 		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugMemoryNoLog|Any CPU.Build.0 = DebugMemoryNoLog|x64
 		{E19C2CBB-86B8-4ED6-ACB2-A35A9BADDA0F}.DebugMemoryNoLog|x64.ActiveCfg = DebugMemoryNoLog|x64
@@ -86,6 +107,12 @@ Global
 		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.Debug|x64.Build.0 = Debug|x64
 		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.Debug|x86.ActiveCfg = Debug|Win32
 		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.Debug|x86.Build.0 = Debug|Win32
+		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugLogNoMemory|Any CPU.ActiveCfg = DebugLogNoMemory|x64
+		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugLogNoMemory|Any CPU.Build.0 = DebugLogNoMemory|x64
+		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugLogNoMemory|x64.ActiveCfg = DebugLogNoMemory|x64
+		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugLogNoMemory|x64.Build.0 = DebugLogNoMemory|x64
+		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugLogNoMemory|x86.ActiveCfg = DebugLogNoMemory|Win32
+		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugLogNoMemory|x86.Build.0 = DebugLogNoMemory|Win32
 		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugMemoryNoLog|Any CPU.ActiveCfg = DebugMemoryNoLog|x64
 		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugMemoryNoLog|Any CPU.Build.0 = DebugMemoryNoLog|x64
 		{CEA17752-5C3A-4E87-A8B7-4AD69A2B499C}.DebugMemoryNoLog|x64.ActiveCfg = DebugMemoryNoLog|x64
@@ -104,6 +131,12 @@ Global
 		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.Debug|x64.Build.0 = Debug|x64
 		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.Debug|x86.ActiveCfg = Debug|Win32
 		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.Debug|x86.Build.0 = Debug|Win32
+		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugLogNoMemory|Any CPU.ActiveCfg = DebugLogNoMemory|x64
+		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugLogNoMemory|Any CPU.Build.0 = DebugLogNoMemory|x64
+		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugLogNoMemory|x64.ActiveCfg = DebugLogNoMemory|x64
+		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugLogNoMemory|x64.Build.0 = DebugLogNoMemory|x64
+		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugLogNoMemory|x86.ActiveCfg = DebugLogNoMemory|Win32
+		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugLogNoMemory|x86.Build.0 = DebugLogNoMemory|Win32
 		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugMemoryNoLog|Any CPU.ActiveCfg = Debug|x64
 		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugMemoryNoLog|Any CPU.Build.0 = Debug|x64
 		{FFBE5670-8AAE-443D-909B-89B7CF403A38}.DebugMemoryNoLog|x64.ActiveCfg = DebugMemoryNoLog|x64
